Address SQLite search review feedback
